+++ src/App.css
@@ -11,7 +11,7 @@
   padding: 0 1rem;
 }
 
-.route-container {
+.toastmsg-container {
   min-height: 3rem;
 }
