+++ src/Components/Playlist/AddToPlaylist.jsx
@@ -1,15 +1,13 @@
 import { useEffect, useState } from "react";
 import { useDataContext } from "../../Context/data-context";
-import { videoExists } from "../Video/videoUtil";
+import { PlaylistModal } from "./PlaylistModal";
 
 export const AddToPlaylist = ({ vid }) => {
   const {
-    state: { playlist },
-    dispatch,
+    state: { playlist }
   } = useDataContext();
 
   const [showModal, setShowModal] = useState(false);
-  const [listName, setListName] = useState("");
 
   useEffect(() => {
     const modal = document.querySelector(".modal-container");
@@ -19,13 +17,6 @@ export const AddToPlaylist = ({ vid }) => {
       }
   }
   },[showModal]);
-
-  const createPlaylist = (event) => {
-    if (event.key === "Enter") {
-      dispatch({ type: "ADD_TO_NEW_PLAYLIST", payload: { vid, listName } });
-      setListName("");
-    }
-  };
 
   const checkInPlaylist = (vid) =>
     playlist.some((item) => item.videos.some((videoId) => videoId === vid));
@@ -49,63 +40,7 @@ export const AddToPlaylist = ({ vid }) => {
           Save{" "}
         </i>
       )}
-      {showModal && (
-        <div className="modal-container">
-          <div className="card modal-content">
-            <h3 className="modal-header">Playlists</h3>
-            <hr />
-            <ul className="list-checkbox">
-              {playlist.map(({ listId, name, videos }) => (
-                <li key={name} className="secondary-txt">
-                  <label>
-                    <input
-                      type="checkbox"
-                      onChange={() =>
-                        dispatch({
-                          type: "TOGGLE_PLAYLIST",
-                          payload: { listId, vid },
-                        })
-                      }
-                      checked={videoExists(videos, vid)}
-                    />{" "}
-                    {name}
-                  </label>
-                </li>
-              ))}
-            </ul>
-            <div className="txt-box">
-              <input
-                onKeyDown={createPlaylist}
-                onChange={(event) => setListName(event.target.value)}
-                value={listName}
-                className="txt-input"
-                type="text"
-                placeholder="Create new playlist"
-              />
-              <span
-                className="txt-icon"
-                onClick={() => {
-                  dispatch({
-                    type: "ADD_TO_NEW_PLAYLIST",
-                    payload: { vid, listName },
-                  });
-                  setListName("");
-                }}
-              >
-                <i className="fas fa-lg fa-plus-circle"></i>
-              </span>
-            </div>
-            <button
-              onClick={() => setShowModal((showModal) => !showModal)}
-              type="button"
-              vid="btn-close-modal"
-              className="btn btn-secondary btn-dismiss"
-            >
-              x
-            </button>
-          </div>
-        </div>
-      )}
+      {showModal && <PlaylistModal vid={vid} setShowModal={setShowModal}/>}
     </>
   );
 };
